pyproject.toml
[tool.poetry]
name = "visuo-synth"
version = "0.1.0"
description = ""
authors = ["Vim Wickramasinghe <vbwickramasinghe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.7"
langchain-anthropic = "^0.3.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
